restructuring server folders
