+++ package.json
@@ -33,6 +33,7 @@
     "babel-loader": "6.2.4",
     "babel-plugin-react-transform": "2.0.2",
     "babel-plugin-transform-es2015-modules-commonjs": "6.7.0",
+    "babel-plugin-transform-object-rest-spread": "6.6.5",
     "babel-plugin-transform-react-constant-elements": "6.5.0",
     "babel-plugin-transform-react-inline-elements": "6.6.5",
     "babel-polyfill": "6.7.2",
