+++ server/web/middleware/html.js
@@ -53,8 +53,8 @@ module.exports = function(){
   return fallback(function(req, res){
     if(!assets) {
       assets = 'production' === process.env.NODE_ENV ? 
-        require('../assets.json') :
-        webpackToAssets(require('../../webpack.config.js'))
+        require('../../assets.json') :
+        webpackToAssets(require('../../../webpack.config.js'))
       
       const css = [], js = []
       Object.keys(assets).forEach(key => {
@@ -116,8 +116,6 @@ module.exports = function(){
       }
     }
 
-    // 2 hour cache, helps prevent flooding and hide dyno deployment downtime
-    // if deployments are breaking, let maintenance:on run for 2h or purge cf cache
     res.set('Cache-Control', 'max-age=60')
 
     res.send(html)
